Create NumberOfSteps.py
You are given two arrays  and . In each step, you can set  if . Determine the minimum number of steps that are required to make all 's equal.

Input format

First line:  
Second line: 
Third line: 
Output format

Print the minimum number of steps that are required to make all 's equal. If it is not possible, then print -1.